refactor logical router routes (#1500)

+++ pkg/controller/gc.go
@@ -599,22 +599,10 @@ func (c *Controller) gcStaticRoute() error {
 		return err
 	}
 	for _, route := range routes {
-		if route.Policy == ovs.PolicyDstIP || route.Policy == "" {
-			if !c.ipam.ContainAddress(route.NextHop) {
-				klog.Infof("gc static route %s %s %s", route.Policy, route.CIDR, route.NextHop)
-				if err := c.ovnClient.DeleteStaticRouteByNextHop(route.NextHop); err != nil {
-					klog.Errorf("failed to delete stale nexthop route %s, %v", route.NextHop, err)
-				}
-			}
-		} else {
-			if strings.Contains(route.CIDR, "/") {
-				continue
-			}
-			if !c.ipam.ContainAddress(route.CIDR) {
-				klog.Infof("gc static route %s %s %s", route.Policy, route.CIDR, route.NextHop)
-				if err := c.ovnClient.DeleteStaticRoute(route.CIDR, c.config.ClusterRouter); err != nil {
-					klog.Errorf("failed to delete stale route %s, %v", route.NextHop, err)
-				}
+		if route.CIDR != "0.0.0.0/0" && route.CIDR != "::/0" && c.ipam.ContainAddress(route.CIDR) {
+			klog.Infof("gc static route %s %s %s", route.Policy, route.CIDR, route.NextHop)
+			if err := c.ovnClient.DeleteStaticRoute(route.CIDR, c.config.ClusterRouter); err != nil {
+				klog.Errorf("failed to delete stale route %s, %v", route.NextHop, err)
 			}
 		}
 	}
